@@ -1,7 +1,8 @@
 {
   "name": "simple-ui",
-  "private": false,
   "version": "0.0.0",
+  "private": false,
+  "type": "module",
   "description": "",
   "main": "dist/cjs/components/index.cjs",
   "module": "dist/es/components/index.mjs",
@@ -27,14 +28,14 @@
   ],
   "scripts": {
     "prepare": "husky install",
-    "dev": "pnpm --filter ./examples run dev",
-    "build": "node --loader ts-node/esm ./packages/vite.config.ts",
+    "dev": "pnpm -C examples run dev",
+    "build": "node --loader ts-node/esm ./packages/vite.config.mts",
     "build:cmp": "pnpm build cmp",
     "build:use": "pnpm build use",
     "build:dir": "pnpm build dir",
     "build:all": "pnpm build cmp use dir",
-    "test": "vitest -c ./packages/vitest.config.ts",
-    "test:cg": "vitest --coverage -c ./packages/vitest.config.ts",
+    "test": "vitest -c ./packages/vitest.config.mts",
+    "test:cg": "vitest --coverage -c ./packages/vitest.config.mts",
     "test:cmp": "pnpm test ./packages/tests/components/*.test.ts",
     "test:use": "pnpm test ./packages/tests/use/*.test.ts"
   },
@@ -42,34 +43,34 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@commitlint/cli": "^17.6.7",
-    "@commitlint/config-conventional": "^17.6.7",
-    "@types/node": "^20.12.5",
-    "@typescript-eslint/eslint-plugin": "^6.1.0",
-    "@typescript-eslint/parser": "^6.1.0",
-    "@vitejs/plugin-vue": "^4.6.2",
-    "@vitest/coverage-v8": "^0.33.0",
-    "@vue/runtime-core": "^3.3.4",
-    "@vue/test-utils": "^2.4.1",
+    "@commitlint/cli": "^19.3.0",
+    "@commitlint/config-conventional": "^19.2.2",
+    "@types/node": "^20.12.12",
+    "@typescript-eslint/eslint-plugin": "^7.10.0",
+    "@typescript-eslint/parser": "^7.10.0",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "@vitest/coverage-v8": "^1.6.0",
+    "@vue/runtime-core": "^3.4.27",
+    "@vue/test-utils": "^2.4.6",
     "@vueuse/core": "^10.9.0",
     "commitizen": "^4.3.0",
     "cz-conventional-changelog": "^3.3.0",
     "eslint": "^8.45.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-vue": "^9.15.1",
-    "husky": "^8.0.3",
-    "jsdom": "^22.1.0",
-    "lint-staged": "^13.2.3",
-    "prettier": "^3.0.0",
-    "sass": "^1.64.1",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-vue": "^9.26.0",
+    "husky": "^9.0.11",
+    "jsdom": "^24.1.0",
+    "lint-staged": "^15.2.5",
+    "prettier": "^3.2.5",
+    "sass": "^1.77.2",
     "ts-node": "^10.9.2",
-    "typescript": "^5.1.6",
-    "vite": "^4.4.6",
-    "vite-plugin-dts": "^3.3.1",
-    "vitepress": "1.0.2",
-    "vitest": "^1.5.0",
-    "vue": "^3.4.3",
-    "vue-router": "^4.2.5",
-    "vue-tsc": "^2.0.13"
+    "typescript": "^5.4.5",
+    "vite": "^5.2.11",
+    "vite-plugin-dts": "^3.9.1",
+    "vitepress": "1.2.2",
+    "vitest": "^1.6.0",
+    "vue": "^3.4.27",
+    "vue-router": "^4.3.2",
+    "vue-tsc": "^2.0.19"
   }
 }
