@@ -11,8 +11,8 @@ function mapFiles(filenames, { lint = true, ts = false } = {}) {
   return commands.map(cmd => (/^vue-tsc/.test(cmd) || /^eslint/.test(cmd) ? cmd : `${cmd} ${filenames.join(" ")}`));
 }
 
-module.exports = {
+export default {
   "*.js": filenames => mapFiles(filenames),
-  "*.{ts,vue}": filenames => mapFiles(filenames, { ts: true }),
+  "*.{ts,mts,vue}": filenames => mapFiles(filenames, { ts: true }),
   "*.{json,md}": filenames => mapFiles(filenames, { lint: false })
 };
